+++ Assets/Scripts/Actor/Props/PropsHealthManager.cs
@@ -5,7 +5,7 @@ namespace Actor.Props
     public class PropsHealthManager : HealthManager
     {
         public bool takesDamage;
-        public override bool TakeDamage(int amount)
+        public override bool TakeDamage(float amount)
         {
             if(takesDamage)
                 return base.TakeDamage(amount);
